requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
pillow==9.1.1
psycopg2>=2.8
# easy-thumbnails==2.8.1 
# easy-thumbnails-rest==1.1.4
django_imagekit==4.1.0
django_cleanup==6.0.0